Fixing DLB memory leaks & bugs; random updates
Cello

  - improved random seeding for parallel runs

Messages

  - BUG FIX: MsgFoo::pack() must delete msg even if buffer is
    non-null
  - bug fix: delete MsgRefine attributes in destructor if not local
  - bug fix: delete field_face_ in ~DataMsg ()
  - removed DataMsg debugging

Performance

  - commented PERF_REDUCE_FOO control_stopping.cpp for load balancing
    due to perf counting errors

+++ src/Enzo/enzo-core/EnzoMsgCheck.hpp
@@ -114,7 +114,7 @@ public: // interface
   }
 
   void set_adapt (const Adapt & adapt) {
-    int size = adapt.data_size();
+    size_t size = adapt.data_size();
     ASSERT2 ("EnzoMsgCheck::set_adapt()",
              "ADAPT_BUFFER_SIZE %d is too small for Adapt object of size %d",
              sizeof(int)*ADAPT_BUFFER_SIZE,size,
